LayoutTests/platform/mac/media/encrypted-media/fps-createMediaKeys-expected.txt

RUN(promise = navigator.requestMediaKeySystemAccess("com.apple.fps", capabilities))
Promise resolved OK
RUN(promise = access.createMediaKeys())
Promise resolved OK
END OF TEST
